requirements.txt
asgiref
Django~=3.1.8
djangorestframework~=3.11.2
Pillow
pytz
six
sqlparse

django-autoslug~=1.9.9